+++ pydoll/extractor/__init__.py
@@ -1,0 +1,16 @@
+from pydoll.extractor.exceptions import (
+    ExtractionException,
+    FieldExtractionFailed,
+    InvalidExtractionModel,
+)
+from pydoll.extractor.field import ExtractionMetadata, Field
+from pydoll.extractor.model import ExtractionModel
+
+__all__ = [
+    'ExtractionException',
+    'ExtractionMetadata',
+    'ExtractionModel',
+    'Field',
+    'FieldExtractionFailed',
+    'InvalidExtractionModel',
+]